+++ tests/testthat/test-basic.R
@@ -49,7 +49,7 @@ test_that(paste0(
 })
 
 test_that(paste0(name(mdl), " fit can be summarized"), {
-  test_summary(est, 41)
+  test_summary(est, 43)
 })
 
 test_that(paste0("Estimates of '", name(mdl), "' are accurate"), {